--- The Quest for the A Plus/src/SchoolRunner.java
@@ -1,11 +1,0 @@
-
-public class SchoolRunner {
-
-	public static void main(String[] args) {
-
-		SchoolSimulator game = new SchoolSimulator();
-		game.setUp();
-
-	}
-
-}
