+++ bin/SchoolSimulator$ChoiceHandler.class
