+++ src/SchoolRunner.java
@@ -1,0 +1,11 @@
+
+public class SchoolRunner {
+
+	public static void main(String[] args) {
+
+		SchoolSimulator game = new SchoolSimulator();
+		game.setUp();
+
+	}
+
+}
